fix/ 비행선 공격력/체력 더블로 변경, 데미지인포도 더블로 변경
추후 적이나 영웅의 스탯도 나중에 바꾸게 되면 takedamage랑 heal 고쳐야함.

+++ Assets/Scripts/Airship/Upgrade/AirshipUpgradeController.cs
@@ -140,7 +140,7 @@ public class AirshipUpgradeController : MonoBehaviour
         return upgradeState.GetLevel(statType);
     }
 
-    public float GetCurrentStat(AirshipStatType statType)
+    public double GetCurrentStat(AirshipStatType statType)
     {
         return statTable.GetStatValue(statType, GetCurrentLevel(statType));
     }
@@ -154,13 +154,13 @@ public class AirshipUpgradeController : MonoBehaviour
         return GetCurrentLevel(statType) + 1;
     }
 
-    public float GetNextStat(AirshipStatType statType)
+    public double GetNextStat(AirshipStatType statType)
     {
         int nextLevel = GetNextLevel(statType);
 
         if (nextLevel < 0)
         {
-            return -1f;
+            return -1d;
         }
 
         return statTable.GetStatValue(statType, nextLevel);
